Make executable a parameter of Converter.new
This is mainly to aid testing. Previously,
ConverterTest#test_executable_not_found was complicated  and threw an
ugly warning. I found a way to silence this, so at least the tests ran
cleanly, but it made the test even more complicated and more tightly
bound to the implementation.

This makes specifying a different executable incredibly clean. It could
even come in useful.